fix delete user token

+++ src/modules/users/infra/typeorm/repositories/UserTokenRepository.ts
@@ -38,7 +38,7 @@ class UserTokenRepository implements IUserTokenRepository {
   }
 
   async deleteById(id: string): Promise<void> {
-    await this.repository.delete(id);
+    await this.repository.delete({ user_id: id });
   }
 
   async findByRefreshToken(
